+++ hospital/templates/hospital/navbar.html
@@ -7,7 +7,7 @@
             <ul class="nav navbar-nav text-light ml-auto">
                 <li class="nav-item" role="presentation"><a class="nav-link" href="{% url 'index' %}">HOME</a></li>
                 
-                <li class="nav-item" role="presentation"><a class="nav-link" href="">PATIENT</a></li>
+                <li class="nav-item" role="presentation"><a class="nav-link" href="{% url 'patient-login' %}">PATIENT</a></li>
                 <li class="nav-item" role="presentation"><a class="nav-link" href="{% url 'doctor-login' %}">DOCTOR</a></li>
                 <li class="nav-item" role="presentation"><a class="nav-link" href="{% url 'admin-login' %}">ADMIN</a></li>
                 <!-- <li class="nav-item" role="presentation"><a class="nav-link" href="">ADMIN LOGIN</a></li> -->
