--- hospital/templates/hospital/patientdashboard.html
@@ -1,6 +1,0 @@
-
-{% block content %}
-<div class="container" style = "margin-top:50px">
-<h2 class="text-center">Patient Dashboard</h2>
-</div>
-{% endblock %}
