+++ patientapp/templates/patientapp/login.html
@@ -1,12 +1,11 @@
-{% extends 'hospital/base.html'  %}
+{% extends 'patientapp/base.html'  %}
 
 {% block content %}
-{% include 'hospital/navbar.html' %}
 <div class="login-dark" >
     <div class="col-md-8 offset-md-2">
         <form class="custom-form" style="margin-top: 20px;" method="post">
             {% csrf_token %}
-            <h2 class="sr-only">Login Form</h2>
+            <h2>Patient Login</h2>
             <div class="illustration"><i class="icon ion-ios-locked-outline"></i></div>
 
             <div class="form-row form-group">
