+++ patientapp/templates/patientapp/patientregistration.html
@@ -1,6 +1,5 @@
-{% extends 'hospital/base.html' %}
+{% extends 'patientapp/base.html' %}
 {% block content %}
-{% include 'hospital/navbar.html' %}
 <div class="row register-form">
     <div class="col-md-8 offset-md-2">
         <form class="custom-form" style="margin-top: 20px;" method="post">
@@ -87,11 +86,9 @@
         
             <button class="btn btn-primary submit-button" type="submit">Create Account</button>
         </form>
-        <hr>
-        <p><a href="{% url 'doctor-registration' %}">Register as Doctor</a></p>
     </div>
 </div>
-{% ifequal error "no" %}
+<!-- {% ifequal error "no" %}
 <script type="text/javascript">
     alert('Patient saved Sucessfully..........')
     window.location = ("{% url 'Login-page' %}")
@@ -102,7 +99,7 @@
 <script type="text/javascript">
     
     alert('Passwords do not match')
-</script>
+</script> -->
 
 {% endifequal%}
 {% endblock %}
